+++ .idea/artifacts/Capatchafy.xml
@@ -2,7 +2,7 @@
   <artifact type="jar" build-on-make="true" name="Capatchafy">
     <output-path>$PROJECT_DIR$/target/</output-path>
     <root id="archive" name="Capatchafy.jar">
-      <element id="module-output" name="Capatchafy" />
+      <element id="module-output" name="Captchafy" />
       <element id="extracted-dir" path="$MAVEN_REPOSITORY$/org/glassfish/grizzly/grizzly-http-server/2.4.0/grizzly-http-server-2.4.0.jar" path-in-jar="/" />
       <element id="extracted-dir" path="$MAVEN_REPOSITORY$/commons-io/commons-io/2.6/commons-io-2.6.jar" path-in-jar="/" />
       <element id="extracted-dir" path="$MAVEN_REPOSITORY$/javax/annotation/javax.annotation-api/1.2/javax.annotation-api-1.2.jar" path-in-jar="/" />
